@@ -5,7 +5,7 @@
   "scripts": {
     "ng": "ng",
     "build": "npm run env -s && ng build --prod",
-    "start": "ng serve --proxy-config proxy.conf.js",
+    "start": "ng serve --aot --proxy-config proxy.conf.js",
     "lint": "ng lint --type-check && stylelint \"src/**/*.scss\" --syntax scss && htmlhint \"src\" --config .htmlhintrc",
     "test": "ng test",
     "test:ci": "npm run lint -s && ng test --single-run --code-coverage",
@@ -22,21 +22,21 @@
     "generate": "ng generate"
   },
   "dependencies": {
-    "@angular/animations": "^4.2.6",
-    "@angular/common": "^4.2.6",
-    "@angular/compiler": "^4.2.6",
-    "@angular/core": "^4.2.6",
-    "@angular/forms": "^4.2.6",
-    "@angular/http": "^4.2.6",
-    "@angular/platform-browser": "^4.2.6",
-    "@angular/platform-browser-dynamic": "^4.2.6",
-    "@angular/router": "^4.2.6",
+    "@angular/animations": "^5.0.0",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.0.0",
+    "@angular/http": "^5.0.0",
+    "@angular/platform-browser": "^5.0.0",
+    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@angular/router": "^5.0.0",
     "@ngx-translate/core": "^8.0.0",
     "@ionic-native/core": "^4.0.0",
     "@ionic-native/keyboard": "^4.0.0",
     "@ionic-native/splash-screen": "^4.0.0",
     "@ionic-native/status-bar": "^4.0.0",
-    "ionic-angular": "^3.5.3",
+    "ionic-angular": "^3.9.2",
     "ionicons": "^3.0.0",
     "core-js": "^2.4.1",
     "lodash": "^4.17.4",
@@ -45,16 +45,16 @@
   },
   "devDependencies": {
     "cordova": "^7.0.0",
-    "@angular/cli": "^1.4.2",
-    "@angular/compiler-cli": "^4.2.6",
-    "@angular/language-service": "^4.2.6",
+    "@angular/cli": "~1.5.0",
+    "@angular/compiler-cli": "^5.0.0",
+    "@angular/language-service": "^5.0.0",
     "@biesbjerg/ngx-translate-extract": "^2.3.1",
     "@ngx-rocket/scripts": "^1.0.4",
     "@types/jasmine": "^2.5.52",
     "@types/jasminewd2": "^2.0.2",
     "@types/lodash": "^4.14.52",
     "@types/node": "^8.0.10",
-    "codelyzer": "~3.2.0",
+    "codelyzer": "~4.0.1",
     "hads": "^1.3.4",
     "htmlhint": "^0.9.13",
     "https-proxy-agent": "^2.0.0",
@@ -66,14 +66,14 @@
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "^1.1.0",
     "karma-junit-reporter": "^1.2.0",
-    "protractor": "~5.1.2",
-    "puppeteer": "^0.11.0",
-    "stylelint": "~8.1.0",
+    "protractor": "~5.2.0",
+    "puppeteer": "^0.12.0",
+    "stylelint": "~8.2.0",
     "stylelint-config-recommended-scss": "~2.0.0",
     "stylelint-config-standard": "~17.0.0",
     "stylelint-scss": "~2.1.0",
-    "ts-node": "^3.2.0",
-    "tslint": "~5.7.0",
+    "ts-node": "~3.3.0",
+    "tslint": "~5.8.0",
     "typescript": "~2.4.2"
   }
 }
