@@ -37,18 +37,22 @@ import 'core-js/es6/set';
 // IE10 and IE11 requires the following for NgClass support on SVG elements
 // import 'classlist.js';  // Run `npm install --save classlist.js`.
 
+// IE10 and IE11 requires the following for the Reflect API.
+// import 'core-js/es6/reflect';
+
+
 // Evergreen browsers require these.
-import 'core-js/es6/reflect';
+// Used for reflect-metadata in JIT. If you use AOT (and only Angular decorators), you can remove.
 import 'core-js/es7/reflect';
 
 
-// Required to support Web Animations `@angular/animation`.
+// Required to support Web Animations `@angular/platform-browser/animations`.
 // Needed for: All but Chrome, Firefox and Opera. http://caniuse.com/#feat=web-animation
 // import 'web-animations-js';  // Run `npm install --save web-animations-js`.
 
 
 // Zone JS is required by Angular itself.
-import 'zone.js/dist/zone';  // Included with Angular-CLI.
+import 'zone.js/dist/zone';  // Included with Angular CLI.
 
 
 /*
